fix local test run
