+++ jest.globalSetup.ts
@@ -1,0 +1,12 @@
+import { MongoMemoryServer } from 'mongodb-memory-server';
+import mongoose from 'mongoose';
+
+let mongoServer: MongoMemoryServer;
+
+export default async function globalSetup() {
+  mongoServer = await MongoMemoryServer.create();
+  const mongoUri = mongoServer.getUri();
+  await mongoose.connect(mongoUri);
+  return { mongoServer };
+}
+export { mongoServer };
