+++ jest.globalTeardown.ts
@@ -1,0 +1,9 @@
+import mongoose from 'mongoose';
+import { mongoServer } from './jest.globalSetup';
+
+export default async function globalTeardown() {
+  await mongoose.disconnect();
+  if (mongoServer) {
+    await mongoServer.stop();
+  }
+}
